Move TimeCell component to separate file
